requirements.txt
Flask==2.0.3
python-dotenv==0.17.1
selenium
datetime
Flask-SQLAlchemy
selenium